[PVR] Refactor CPVRGUIActions step 4: Move playback related actions to own class.
